requirements.txt
autogen
openai
pydantic>=2.0
python-dotenv
jsonschema